Big Update
Just a lot of changes.

+++ Assets/Scripts/Other/SaveSystem.cs
@@ -1,13 +1,19 @@
 using System.Collections;
 using System.Collections.Generic;
+using Unity.VisualScripting;
 using UnityEngine;
 
 public static class SaveSystem
 {
-    private static string path = Application.persistentDataPath;
-    private static string oldPath = Application.dataPath + "/SavedLevels";
-    public static void saveLevel(LevelData levelData, string levelName)
+    private static string path = Application.persistentDataPath + "/SavedLevels";
+    private static string streamingPath = Application.streamingAssetsPath;
+    private static string storyPath = Application.persistentDataPath + "/SavedStories";
+    
+    public static void saveLevel(LevelData levelData)
     {
+        string levelName = levelData.levelName;
+        createFolder();
+
         // Convert level data from LevelData object to string
         string levelDataString = JsonUtility.ToJson(levelData);
 
@@ -16,11 +22,14 @@ public static class SaveSystem
     }
     public static void deleteLevel(string levelName)
     {
-        // Delete level data file
-        System.IO.File.Delete(path + "/" + levelName + ".json");
+        if (exists(levelName))
+        {
+            System.IO.File.Delete(path + "/" + levelName + ".json");
+        }
     }
-    public static LevelData getLevel(string levelName)
+    public static LevelData getLevel(string levelName, bool story)
     {
+        string path = story ? streamingPath : SaveSystem.path;
         if (!System.IO.File.Exists(path + "/" + levelName + ".json"))
         {
             return null;
@@ -36,6 +45,8 @@ public static class SaveSystem
     }
     public static string[] getAllNames()
     {
+        createFolder();
+
         // Get all files with .json extension in SavedLevels folder
         string[] filePaths = System.IO.Directory.GetFiles(path + "/", "*.json");
 
@@ -51,5 +62,77 @@ public static class SaveSystem
     public static bool exists(string levelName)
     {
         return System.IO.File.Exists(path + "/" + levelName + ".json");
+    }
+
+    public static void saveStory(StoryData storyData)
+    {
+        string storyName = storyData.SaveName;
+        createFolder();
+
+        // Convert story data from StoryData object to string
+        string storyDataString = JsonUtility.ToJson(storyData);
+
+        // Save story data to file
+        System.IO.File.WriteAllText(storyPath + "/" + storyName + ".json", storyDataString);
+    }
+    public static void deleteStory(string storyName)
+    {
+        if (existsStory(storyName))
+        {
+            System.IO.File.Delete(storyPath + "/" + storyName + ".json");
+        }
+    }
+    public static StoryData getStory(string storyName)
+    {
+        if (!System.IO.File.Exists(storyPath + "/" + storyName + ".json"))
+        {
+            return null;
+        }
+
+        // Get story data from file
+        string storyDataString = System.IO.File.ReadAllText(storyPath + "/" + storyName + ".json");
+
+        // Convert story data from string to StoryData object
+        StoryData storyData = JsonUtility.FromJson<StoryData>(storyDataString);
+
+        return storyData;
+    }
+    public static string[] getAllStoryNames()
+    {
+        createFolder();
+
+        // Get all files with .json extension in SavedStories folder
+        string[] filePaths = System.IO.Directory.GetFiles(storyPath + "/", "*.json");
+
+        // Get story names from file paths
+        string[] storyNames = new string[filePaths.Length];
+        for (int i = 0; i < filePaths.Length; i++)
+        {
+            storyNames[i] = filePaths[i].Substring(filePaths[i].LastIndexOf('/') + 1,
+                                                                  filePaths[i].LastIndexOf('.') - filePaths[i].LastIndexOf('/') - 1);
+        }
+        return storyNames;
+    }
+    public static bool existsStory(string storyName)
+    {
+        return System.IO.File.Exists(storyPath + "/" + storyName + ".json");
+    }
+    public static bool existsStoryLevel(string storyLevel)
+    {
+        return System.IO.File.Exists(streamingPath + "/" + storyLevel + ".json");
+    }
+
+    public static void createFolder()
+    {
+        // Create SavedLevels folder if it doesn't exist
+        if (!System.IO.Directory.Exists(path))
+        {
+            System.IO.Directory.CreateDirectory(path);
+        }
+        // Create SavedStories folder if it doesn't exist
+        if (!System.IO.Directory.Exists(storyPath))
+        {
+            System.IO.Directory.CreateDirectory(storyPath);
+        }
     }
 }
